List of Content functions

Every Content subclass has the following built-in functions:

Base.length

Base.size (1-tuple of length)

Base.firstindex, Base.lastindex(1-based or inherited from its index)

Base.getindex select by Int(single item), UnitRange{Int}(slice), and Symbol(record field)

Base.iterate

Base.:(==) (equality defined by values: a ListOffsetArray and a ListArray may be considered the same)

Base.push!

Base.append!

Base.show

They also have the following functions for manipulating and checking structure:

AwkwardArray.parameters_of gets all parameters

AwkwardArray.has_parameter returns true if a parameter exists

AwkwardArray.get_parameter returns a parameter or raises an error

AwkwardArray.with_parameter returns a copy of this node with a specified parameter

AwkwardArray.copy shallow-copy of the array, allowing properties to be replaced

AwkwardArray.is_valid verifies that the structure adheres to Awkward Array's protocol

They have the following functions for filling an array:

AwkwardArray.end_list!: closes off a ListType array (ListOffsetArray, ListArray, or RegularArray) in the manner of Python's ak.ArrayBuilder (no begin_list is necessary)

AwkwardArray.end_record! closes off a RecordArray

AwkwardArray.end_tuple! closes off a TupleArray

AwkwardArray.push_null! pushes a missing value onto OptionType arrays (IndexedOptionArray](@ref) ByteMaskedArray BitMaskedArray or UnmaskedArray)

AwkwardArray.push_dummy! pushes an unspecified value onto the array (used by ByteMaskedArray and BitMaskedArray which need to have a placeholder in memory behind each missing value)

RecordArrayand TupleArray have the following for selecting fields (as opposed to rows):

AwkwardArray.slot gets a RecordArrayor TupleArray field, to avoid conflicts with Base.getindex for TupleArrays (both use integers to select a field)

AwkwardArray.Record scalar representation of an item from a RecordArray

AwkwardArray.SlotRecord scalar representation of an item from a TupleArray(note: not the same as Base.Tuple)

UnionArrayhas the following for dealing with specializations:

AwkwardArray.Specialization selects a UnionArrayspecialization for push! append! etc.

Finally, all Contentsubclasses can be converted with the following:

AwkwardArray.layout_for returns an appropriately-nested Contenttype for a given Julia type (DataType)

AwkwardArray.from_iter converts Julia data into an Awkward Array

AwkwardArray.to_vector converts an Awkward Array into Julia data

AwkwardArray.from_buffers constructs an Awkward Array from a Form (JSON), length, and buffers for zero-copy passing from Python

AwkwardArray.to_buffers deconstructs an Awkward Array into a Form (JSON), length, and buffers for zero-copy passing to Python

Array functions

AwkwardArray.ByteStringArray — Method

ByteStringArray(;
    parameters::Parameters = Parameters(),
    byte_parameters::Parameters = Parameters(),
)

AwkwardArray.ByteStringArray — Method

ByteStringArray(
    starts::INDEX,
    stops::INDEX,
    data::AbstractVector{UInt8};
    parameters::Parameters = Parameters(),
    byte_parameters::Parameters = Parameters(),
) where {INDEX<:IndexBig}

AwkwardArray.ByteStringOffsetArray — Method

ByteStringOffsetArray(;
    parameters::Parameters = Parameters(),
    byte_parameters::Parameters = Parameters(),
)

AwkwardArray.ByteStringOffsetArray — Method

ByteStringOffsetArray(
    offsets::INDEX,
    data::AbstractVector{UInt8};
    parameters::Parameters = Parameters(),
    byte_parameters::Parameters = Parameters(),
) where {INDEX<:IndexBig}

ListType with behavior = :bytestring.

AwkwardArray.ByteStringRegularArray — Method

ByteStringRegularArray(
    data::AbstractVector{UInt8},
    size::Int;
    zeros_length::Int = 0,
    parameters::Parameters = Parameters(),
    byte_parameters::Parameters = Parameters(),
)

AwkwardArray.ByteStringRegularArray — Method

ByteStringRegularArray(
    size::Int;
    parameters::Parameters = Parameters(),
    byte_parameters::Parameters = Parameters(),
)

AwkwardArray.ByteStringRegularArray — Method

ByteStringRegularArray(;
    parameters::Parameters = Parameters(),
    byte_parameters::Parameters = Parameters(),
)

AwkwardArray.StringArray — Method

StringArray(;
    parameters::Parameters = Parameters(),
    char_parameters::Parameters = Parameters(),
)

AwkwardArray.StringArray — Method

StringArray(
    starts::INDEX,
    stops::INDEX,
    data::AbstractVector{UInt8};
    parameters::Parameters = Parameters(),
    char_parameters::Parameters = Parameters(),
) where {INDEX<:IndexBig}

AwkwardArray.StringArray — Method

StringArray(
    starts::INDEX,
    stops::INDEX,
    data::String;   # data provided as a String, rather than AbstractVector{UInt8}
    parameters::Parameters = Parameters(),
    char_parameters::Parameters = Parameters(),
) where {INDEX<:IndexBig}

AwkwardArray.StringOffsetArray — Method

StringOffsetArray(; parameters = Parameters(), char_parameters = Parameters())

AwkwardArray.StringOffsetArray — Method

StringOffsetArray(
    offsets::INDEX,
    data::AbstractVector{UInt8};
    parameters::Parameters = Parameters(),
    char_parameters::Parameters = Parameters(),
) where {INDEX<:IndexBig}

AwkwardArray.StringOffsetArray — Method

StringOffsetArray(
    offsets::INDEX,
    data::String;
    parameters::Parameters = Parameters(),
    char_parameters::Parameters = Parameters(),
) where {INDEX<:IndexBig}

AwkwardArray.StringRegularArray — Method

StringRegularArray(
    data::AbstractVector{UInt8},
    size::Int;
    zeros_length::Int = 0,
    parameters::Parameters = Parameters(),
    char_parameters::Parameters = Parameters(),
)

AwkwardArray.StringRegularArray — Method

StringRegularArray(
    size::Int;
    parameters::Parameters = Parameters(),
    char_parameters::Parameters = Parameters(),
)

AwkwardArray.StringRegularArray — Method

StringRegularArray(
    data::String,
    size::Int;
    zeros_length::Int = 0,
    parameters::Parameters = Parameters(),
    char_parameters::Parameters = Parameters(),
)

AwkwardArray.StringRegularArray — Method

StringRegularArray(;
    parameters::Parameters = Parameters(),
    char_parameters::Parameters = Parameters(),
)

AwkwardArray._alternate — Method

_alternate(range::AbstractRange{Int64})

AwkwardArray._get_buffer — Method

_get_buffer(
    form_key::Union{Nothing,String},
    attribute::String,
    buffer_key::BUFFER_KEY_FUNCTION,
    containers::Dict{String,BUFFER},
) where {BUFFER<:AbstractVector{UInt8},BUFFER_KEY_FUNCTION<:Function}

AwkwardArray._get_index — Method

_get_index(
    form_snippet::String,
    length::Int64,
    buffer::BUFFER,
) where {BUFFER<:AbstractVector{UInt8}}

AwkwardArray._horizontal — Method

_horizontal(data::Any, limit_cols::Int)

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::EmptyArray,
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
)

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::PrimitiveArray{ITEM,BUFFER},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {ITEM,BUFFER<:AbstractVector{ITEM}}

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::TupleArray{CONTENTS},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {CONTENTS<:Base.Tuple{Vararg{Content}}}

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::RecordArray{FIELDS,CONTENTS},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {FIELDS,CONTENTS<:Base.Tuple{Vararg{Content}}}

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::UnionArray{TAGS,INDEX,CONTENTS},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {TAGS<:Index8,INDEX<:IndexBig,CONTENTS<:Base.Tuple{Vararg{Content}}}

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::BitMaskedArray{CONTENT},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {CONTENT<:Content}

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::RegularArray{CONTENT},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {CONTENT<:Content}

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::UnmaskedArray{CONTENT},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {CONTENT<:Content}

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::ByteMaskedArray{INDEX,CONTENT},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {INDEX<:IndexBool,CONTENT<:Content}

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::IndexedArray{INDEX,CONTENT},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {INDEX<:IndexBig,CONTENT<:Content}

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::IndexedOptionArray{INDEX,CONTENT},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {INDEX<:IndexBigSigned,CONTENT<:Content}

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::ListArray{INDEX,CONTENT},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {INDEX<:IndexBig,CONTENT<:Content}

AwkwardArray._to_buffers! — Method

_to_buffers!(
    layout::ListOffsetArray{INDEX,CONTENT},
    number::Vector{Int64},
    containers::Dict{String,AbstractVector{UInt8}},
) where {INDEX<:IndexBig,CONTENT<:Content}

AwkwardArray._to_buffers_index — Method

_to_buffers_index(IndexType::DataType)

AwkwardArray._to_buffers_parameters — Method

_to_buffers_parameters(layout::CONTENT) where {BEHAVIOR,CONTENT<:Content{BEHAVIOR}}

AwkwardArray._vertical — Method

_vertical(data::Union{Content,Record,Tuple}, limit_rows::Int, limit_cols::Int)

AwkwardArray.check_primitive_type — Method

check_primitive_type(ITEM)

A utility that provides a string representation for various primitive types in Julia. It helps in identifying the type of an item and mapping it to a human-readable format. This is useful in scenarios where type information needs to be logged, serialized, or displayed. If a type is not recognized, it defaults to returning unknown.

AwkwardArray.compatible — Method

compatible(parameters1::Parameters, parameters2::Parameters)

AwkwardArray.copy — Function

copy(behavior::Union{Unset,Symbol} = Unset())

AwkwardArray.copy — Method

copy(
    layout::BitMaskedArray{CONTENT1,BEHAVIOR};
    mask::Union{Unset,BitVector} = Unset(),
    content::Union{Unset,CONTENT2} = Unset(),
    valid_when::Union{Unset,Bool} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
) where {CONTENT1<:Content,CONTENT2<:Content,BEHAVIOR}

AwkwardArray.copy — Method

copy(
    layout::ByteMaskedArray{INDEX1,CONTENT1,BEHAVIOR};
    mask::Union{Unset,INDEX2} = Unset(),
    content::Union{Unset,CONTENT2} = Unset(),
    valid_when::Union{Unset,Bool} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
) where {INDEX1<:IndexBool,INDEX2<:IndexBool,CONTENT1<:Content,CONTENT2<:Content,BEHAVIOR}

AwkwardArray.copy — Method

copy(
    layout::IndexedArray{INDEX1,CONTENT1,BEHAVIOR};
    index::Union{Unset,INDEX2} = Unset(),
    content::Union{Unset,CONTENT2} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
) where {INDEX1<:IndexBig,INDEX2<:IndexBig,CONTENT1<:Content,CONTENT2<:Content,BEHAVIOR}

AwkwardArray.copy — Method

copy(
    layout::IndexedOptionArray{INDEX1,CONTENT1,BEHAVIOR};
    index::Union{Unset,INDEX2} = Unset(),
    content::Union{Unset,CONTENT2} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
) where {
    INDEX1<:IndexBigSigned,
    INDEX2<:IndexBigSigned,
    CONTENT1<:Content,
    CONTENT2<:Content,
    BEHAVIOR,
}

Copy of an IndexedOptionArray, potentially with some modifications to its fields.

AwkwardArray.copy — Method

copy(
    layout::ListArray{INDEX1,CONTENT1,BEHAVIOR};
    starts::Union{Unset,INDEX2} = Unset(),
    stops::Union{Unset,INDEX2} = Unset(),
    content::Union{Unset,CONTENT2} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
) where {INDEX1<:IndexBig,INDEX2<:IndexBig,CONTENT1<:Content,CONTENT2<:Content,BEHAVIOR}

Copy of a ListArray.

AwkwardArray.copy — Method

copy(
    layout::ListOffsetArray{INDEX1,CONTENT1,BEHAVIOR};
    offsets::Union{Unset,INDEX2} = Unset(),
    content::Union{Unset,CONTENT2} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
) where {INDEX1<:IndexBig,INDEX2<:IndexBig,CONTENT1<:Content,CONTENT2<:Content,BEHAVIOR}

Copy of a ListOffsetArray.

AwkwardArray.copy — Method

copy(
    layout::PrimitiveArray{ITEM,BUFFER1,BEHAVIOR};
    data::Union{Unset,BUFFER2} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
) where {ITEM,BUFFER1<:AbstractVector{ITEM},BUFFER2<:AbstractVector,BEHAVIOR}

AwkwardArray.copy — Method

copy(
    layout::RecordArray{FIELDS1,CONTENTS1,BEHAVIOR};
    contents::Union{Unset,NamedTuple{FIELDS2,CONTENTS2}} = Unset(),
    length::Union{Unset,Int64} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
) where {
    FIELDS1,
    FIELDS2,
    CONTENTS1<:Base.Tuple{Vararg{Content}},
    CONTENTS2<:Base.Tuple{Vararg{Content}},
    BEHAVIOR,
}

AwkwardArray.copy — Method

copy(
    layout::RegularArray{CONTENT1,BEHAVIOR};
    content::Union{Unset,CONTENT2} = Unset(),
    size::Union{Unset,Int} = Unset(),
    zeros_length::Union{Unset,Int} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
) where {CONTENT1<:Content,CONTENT2<:Content,BEHAVIOR}

Copy of a RegularArray.

AwkwardArray.copy — Method

copy(
    layout::TupleArray{CONTENTS1,BEHAVIOR};
    contents::Union{Unset,CONTENTS2} = Unset(),
    length::Union{Unset,Int64} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
) where {
    CONTENTS1<:Base.Tuple{Vararg{Content}},
    CONTENTS2<:Base.Tuple{Vararg{Content}},
    BEHAVIOR,
}

AwkwardArray.copy — Method

copy(
    layout::UnionArray{TAGS1,INDEX1,CONTENTS1,BEHAVIOR};
    tags::Union{Unset,TAGS2} = Unset(),
    index::Union{Unset,INDEX2} = Unset(),
    contents::Union{Unset,CONTENTS2} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
)

AwkwardArray.copy — Method

copy(
    layout::UnmaskedArray{CONTENT1,BEHAVIOR};
    content::Union{Unset,CONTENT2} = Unset(),
    parameters::Union{Unset,Parameters} = Unset(),
    behavior::Union{Unset,Symbol} = Unset(),
) where {CONTENT1<:Content,CONTENT2<:Content,BEHAVIOR}

AwkwardArray.default_buffer_key — Method

default_buffer_key(form_key::String, attribute::String)

AwkwardArray.end_list! — Method

end_list!(layout::BitMaskedArray)

AwkwardArray.end_list! — Method

end_list!(layout::ByteMaskedArray)

AwkwardArray.end_list! — Method

end_list!(layout::IndexedArray)

AwkwardArray.end_list! — Method

end_list!(layout::IndexedOptionArray)

AwkwardArray.end_list! — Method

end_list!(layout::ListArray)

AwkwardArray.end_list! — Method

end_list!(layout::ListOffsetArray)

AwkwardArray.end_list! — Method

end_list!(layout::RegularArray)

AwkwardArray.end_list! — Method

end_list!(special::Specialization)

AwkwardArray.end_list! — Method

end_list!(layout::UnmaskedArray)

AwkwardArray.end_record! — Method

end_record!(layout::BitMaskedArray)

AwkwardArray.end_record! — Method

end_record!(layout::ByteMaskedArray)

AwkwardArray.end_record! — Method

end_record!(layout::IndexedArray)

AwkwardArray.end_record! — Method

end_record!(layout::IndexedOptionArray)

AwkwardArray.end_record! — Method

end_record!(layout::RecordArray)

AwkwardArray.end_record! — Method

end_record!(special::Specialization)

AwkwardArray.end_record! — Method

end_record!(layout::UnmaskedArray)

AwkwardArray.end_tuple! — Method

end_tuple!(layout::BitMaskedArray)

AwkwardArray.end_tuple! — Method

end_tuple!(layout::ByteMaskedArray)

AwkwardArray.end_tuple! — Method

end_tuple!(layout::IndexedArray)

AwkwardArray.end_tuple! — Method

end_tuple!(layout::IndexedOptionArray)

AwkwardArray.end_tuple! — Method

end_tuple!(special::Specialization)

AwkwardArray.end_tuple! — Method

end_tuple!(layout::TupleArray)

AwkwardArray.end_tuple! — Method

end_tuple!(layout::UnmaskedArray)

AwkwardArray.from_buffers — Method

from_buffers(
    form::Dict{String,Any},
    length::Int,
    containers::Dict{String,BUFFER};
    buffer_key::BUFFER_KEY_FUNCTION = default_buffer_key,
) where {BUFFER<:AbstractVector{UInt8},BUFFER_KEY_FUNCTION<:Function}

AwkwardArray.from_buffers — Method

from_buffers(
    form::String,
    length::Int,
    containers::Dict{String,BUFFER};
    buffer_key::BUFFER_KEY_FUNCTION = default_buffer_key,
) where {BUFFER<:AbstractVector{UInt8},BUFFER_KEY_FUNCTION<:Function}

AwkwardArray.from_iter — Method

from_iter(input)

AwkwardArray.get_parameter — Method

get_parameter(content::EmptyArray, key::String)

AwkwardArray.get_parameter — Method

get_parameter(parameters::Parameters, key::String)

AwkwardArray.get_parameter — Method

get_parameter(content::CONTENT, key::String) where {CONTENT<:Content}

Return a parameter or raises an error.

AwkwardArray.has_parameter — Method

has_parameter(content::EmptyArray, key::String)

AwkwardArray.has_parameter — Method

has_parameter(parameters::Parameters, key::String)

AwkwardArray.has_parameter — Method

has_parameter(content::CONTENT, key::String) where {CONTENT<:Content}

Return true if a parameter exists.

AwkwardArray.is_valid — Method

is_valid(layout::BitMaskedArray)

AwkwardArray.is_valid — Method

is_valid(layout::ByteMaskedArray)

AwkwardArray.is_valid — Method

is_valid(layout::EmptyArray)

AwkwardArray.is_valid — Method

is_valid(layout::IndexedArray)

AwkwardArray.is_valid — Method

is_valid(layout::IndexedOptionArray)

AwkwardArray.is_valid — Method

is_valid(layout::ListArray)

Check if a ListArray is valid.

AwkwardArray.is_valid — Method

is_valid(layout::ListOffsetArray)

Check if a ListOffsetArray is valid.

AwkwardArray.is_valid — Method

is_valid(layout::PrimitiveArray)

AwkwardArray.is_valid — Method

is_valid(layout::RecordArray)

AwkwardArray.is_valid — Method

is_valid(layout::RegularArray)

Check if a RegularArray is valid.

AwkwardArray.is_valid — Method

is_valid(layout::TupleArray)

AwkwardArray.is_valid — Method

is_valid(layout::UnionArray)

AwkwardArray.is_valid — Method

is_valid(layout::UnmaskedArray)

AwkwardArray.isprimitive — Method

isprimitive(t::Type)

Determine if a given type is one of the fundamental data types in Julia that are typically considered primitive. These include:

Integer: Represents all integer types (e.g., Int64, UInt8). AbstractFloat: Represents all floating-point types (e.g., Float64, Float32). Bool: Represents boolean values (true and false). Char: Represents character values.

AwkwardArray.layout_for — Method

layout_for(ItemType)

AwkwardArray.parameters_of — Method

parameters_of(content::EmptyArray)

AwkwardArray.parameters_of — Method

parameters_of(content::CONTENT) where {CONTENT<:Content}

Return a list of all parameters.

AwkwardArray.push_dummy! — Method

push_dummy!(layout::BitMaskedArray)

AwkwardArray.push_dummy! — Method

push_dummy!(layout::ByteMaskedArray)

AwkwardArray.push_dummy! — Method

push_dummy!(layout::IndexedArray)

AwkwardArray.push_dummy! — Method

push_dummy!(layout::IndexedOptionArray)

AwkwardArray.push_dummy! — Method

push_dummy!(layout::ListArray)

AwkwardArray.push_dummy! — Method

push_dummy!(layout::ListOffsetArray)

AwkwardArray.push_dummy! — Method

push_dummy!(layout::RecordArray)

AwkwardArray.push_dummy! — Method

push_dummy!(layout::RegularArray)

AwkwardArray.push_dummy! — Method

push_dummy!(special::Specialization)

AwkwardArray.push_dummy! — Method

push_dummy!(layout::TupleArray)

AwkwardArray.push_dummy! — Method

push_dummy!(layout::UnmaskedArray)

AwkwardArray.push_dummy! — Method

push_dummy!(layout::PrimitiveArray{ITEM}) where {ITEM}

AwkwardArray.push_null! — Method

push_null!(layout::BitMaskedArray)

AwkwardArray.push_null! — Method

push_null!(layout::ByteMaskedArray)

AwkwardArray.push_null! — Method

push_null!(layout::IndexedOptionArray)

AwkwardArray.push_null! — Method

push_null!(
    special::Specialization{ARRAY,TAGGED},
) where {ARRAY<:UnionArray,TAGGED<:OptionType}

AwkwardArray.slot — Method

slot(
    layout::TupleArray{CONTENTS,BEHAVIOR},
    f::Int,
) where {CONTENTS<:Base.Tuple{Vararg{Content}},BEHAVIOR}

AwkwardArray.slot — Method

slot(
    layout::RecordArray{FIELDS,CONTENTS,BEHAVIOR},
    f::Symbol,
) where {FIELDS,CONTENTS<:Base.Tuple{Vararg{Content}},BEHAVIOR}

AwkwardArray.to_buffers — Method

to_buffers(layout::Content)

AwkwardArray.to_vector — Method

to_vector(layout::Content; view::Bool = false, na::Union{Missing,Nothing} = missing)

AwkwardArray.to_vector — Method

to_vector(
    layout::EmptyArray,
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
)

AwkwardArray.to_vector — Method

to_vector(
    record::Record{FIELDS,CONTENTS};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {FIELDS,CONTENTS<:Base.Tuple{Vararg{Content}}} = NamedTuple{FIELDS}

AwkwardArray.to_vector — Method

to_vector(
    layout::RegularArray{PrimitiveArray{UInt8,BUFFER,:char},:string},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {BUFFER<:AbstractVector{UInt8}}

AwkwardArray.to_vector — Method

to_vector(
    layout::ListArray{INDEX,PrimitiveArray{UInt8,BUFFER,:char},:string},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {INDEX<:IndexBig,BUFFER<:AbstractVector{UInt8}}

AwkwardArray.to_vector — Method

to_vector(
    layout::ListOffsetArray{INDEX,PrimitiveArray{UInt8,BUFFER,:char},:string},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {INDEX<:IndexBig,BUFFER<:AbstractVector{UInt8}}

AwkwardArray.to_vector — Method

to_vector(
    layout::TupleArray{CONTENTS},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {CONTENTS<:Base.Tuple{Vararg{Content}}}

AwkwardArray.to_vector — Method

to_vector(
    layout::RecordArray{FIELDS,CONTENTS},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {FIELDS,CONTENTS<:Base.Tuple{Vararg{Content}}}

AwkwardArray.to_vector — Method

to_vector(
    layout::UnionArray{TAGS,INDEX,CONTENTS},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {TAGS<:Index8,INDEX<:IndexBig,CONTENTS<:Base.Tuple}

AwkwardArray.to_vector — Method

to_vector(
    layout::BitMaskedArray{CONTENT},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {CONTENT<:Content}

AwkwardArray.to_vector — Method

to_vector(
    layout::RegularArray{CONTENT},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {CONTENT<:Content}

AwkwardArray.to_vector — Method

to_vector(
    layout::UnmaskedArray{CONTENT},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {CONTENT<:Content}

AwkwardArray.to_vector — Method

to_vector(
    layout::ByteMaskedArray{INDEX,CONTENT},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {INDEX<:IndexBool,CONTENT<:Content}

AwkwardArray.to_vector — Method

to_vector(
    layout::IndexedArray{INDEX,CONTENT},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {INDEX<:IndexBig,CONTENT<:Content}

AwkwardArray.to_vector — Method

to_vector(
    layout::IndexedOptionArray{INDEX,CONTENT},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {INDEX<:IndexBig,CONTENT<:Content}

AwkwardArray.to_vector — Method

to_vector(
    layout::ListArray{INDEX,CONTENT},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {INDEX<:IndexBig,CONTENT<:Content}

AwkwardArray.to_vector — Method

to_vector(
    layout::ListOffsetArray{INDEX,CONTENT},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {INDEX<:IndexBig,CONTENT<:Content}

AwkwardArray.to_vector — Method

to_vector(
    layout::PrimitiveArray{ITEM},
    r::UnitRange{Int};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {ITEM}

AwkwardArray.to_vector — Method

to_vector(
    tuple::Tuple{CONTENTS};
    view::Bool = false,
    na::Union{Missing,Nothing} = missing,
) where {CONTENTS<:Base.Tuple{Vararg{Content}}}

AwkwardArray.to_vector_or_scalar — Method

to_vector_or_scalar(x; view::Bool = false, na = missing)

AwkwardArray.to_vector_or_scalar — Method

to_vector_or_scalar(x::Content; view::Bool = false, na::Union{Missing,Nothing} = missing)

AwkwardArray.with_parameter — Method

with_parameter(parameters::Parameters, pair::Pair{String,String})

AwkwardArray.with_parameter — Method

with_parameter(parameters::Parameters, pair::Pair{String,<:Any})

Base.:== — Method

Base.:(==)(layout1::Content, layout2::Content)

Two Content objects are considered equal only if they have the same elements in the same order.

Base.:== — Method

Base.:(==)(
    layout1::RecordArray{FIELDS,CONTENTS1},
    layout2::RecordArray{FIELDS,CONTENTS2},
) where {
    FIELDS,
    CONTENTS1<:Base.Tuple{Vararg{Content}},
    CONTENTS2<:Base.Tuple{Vararg{Content}},
}

Base.:== — Method

Base.:(==)(
    layout1::Record{FIELDS,CONTENTS1},
    layout2::Record{FIELDS,CONTENTS2},
) where {
    FIELDS,
    CONTENTS1<:Base.Tuple{Vararg{Content}},
    CONTENTS2<:Base.Tuple{Vararg{Content}},
}

Base.:== — Method

Base.:(==)(
    layout1::SlotRecord{CONTENTS1},
    layout2::SlotRecord{CONTENTS2},
) where {
    N,
    CONTENTS1<:Base.Tuple{Vararg{Content,N}},
    CONTENTS2<:Base.Tuple{Vararg{Content,N}},
}

Base.:== — Method

Base.:(==)(
    layout1::TupleArray{CONTENTS1},
    layout2::TupleArray{CONTENTS2},
) where {
    N,
    CONTENTS1<:Base.Tuple{Vararg{Content,N}},
    CONTENTS2<:Base.Tuple{Vararg{Content,N}},
}

Base.append! — Method

Base.append!(layout::Content, input)

Append multiple elements (from another collection) to an instance of Content.

Base.append! — Method

Base.append!(layout::OptionType, input)

Appending elements to an array of type OptionType, handling missing values.

Base.append! — Method

Base.append!(special::Specialization, input)

Base.eltype — Method

Base.eltype(layout::BitMaskedArray)

Base.eltype — Method

Base.eltype(layout::ByteMaskedArray)

Base.eltype — Method

Base.eltype(layout::EmptyArray)

Base.eltype — Method

Base.eltype(layout::IndexedArray)

Base.eltype — Method

Base.eltype(layout::IndexedOptionArray)

Base.eltype — Method

Base.eltype(layout::ListType)

Base.eltype — Method

Base.eltype(layout::PrimitiveArray)

Base.eltype — Method

Base.eltype(layout::UnionArray)

Base.eltype — Method

Base.eltype(layout::UnmaskedArray)

Base.firstindex — Method

Base.firstindex(layout::BitMaskedArray)

Base.firstindex — Method

Base.firstindex(layout::ByteMaskedArray)

Base.firstindex — Method

Base.firstindex(layout::EmptyArray)

Base.firstindex — Method

Base.firstindex(layout::IndexedArray)

Base.firstindex — Method

Base.firstindex(layout::IndexedOptionArray)

Base.firstindex — Method

Base.firstindex(layout::ListArray)

First index of a ListArray starts.

Base.firstindex — Method

Base.firstindex(layout::ListOffsetArray)

First index of a ListOffsetArray offsets.

Base.firstindex — Method

Base.firstindex(layout::PrimitiveArray)

Base.firstindex — Method

Base.firstindex(layout::RecordArray)

Base.firstindex — Method

Base.firstindex(layout::RegularArray)

First index of a RegularArray is always 1.

Base.firstindex — Method

Base.firstindex(layout::TupleArray)

Base.firstindex — Method

Base.firstindex(layout::UnionArray)

Base.firstindex — Method

Base.firstindex(layout::UnmaskedArray)

Base.getindex — Method

Base.getindex(layout::BitMaskedArray, i::Int)

Base.getindex — Method

Base.getindex(layout::BitMaskedArray, f::Symbol)

Base.getindex — Method

Base.getindex(layout::BitMaskedArray, r::UnitRange{Int})

Base.getindex — Method

Base.getindex(layout::ByteMaskedArray, i::Int)

Base.getindex — Method

Base.getindex(layout::ByteMaskedArray, f::Symbol)

Base.getindex — Method

Base.getindex(layout::ByteMaskedArray, r::UnitRange{Int})

Base.getindex — Method

Base.getindex(layout::EmptyArray, i::Int)

Base.getindex — Method

Base.getindex(layout::EmptyArray, r::UnitRange{Int})

Base.getindex — Method

Base.getindex(layout::IndexedArray, i::Int)

Base.getindex — Method

Base.getindex(layout::IndexedArray, f::Symbol)

Base.getindex — Method

Base.getindex(layout::IndexedArray, r::UnitRange{Int})

Base.getindex — Method

Base.getindex(layout::IndexedOptionArray, i::Int)

Base.getindex — Method

Base.getindex(layout::IndexedOptionArray, f::Symbol)

Base.getindex — Method

Base.getindex(layout::IndexedOptionArray, r::UnitRange{Int})

Base.getindex — Method

Base.getindex(layout::ListArray, i::Int)

Base.getindex — Method

Base.getindex(layout::ListArray, f::Symbol)

Base.getindex — Method

Base.getindex(layout::ListArray, r::UnitRange{Int})

Base.getindex — Method

Base.getindex(layout::ListOffsetArray, i::Int)

Base.getindex — Method

Base.getindex(layout::ListOffsetArray, f::Symbol)

Base.getindex — Method

Base.getindex(layout::ListOffsetArray, r::UnitRange{Int})

Base.getindex — Method

Base.getindex(layout::PrimitiveArray, i::Int)

Base.getindex — Method

Base.getindex(layout::PrimitiveArray, r::UnitRange{Int})

Base.getindex — Method

Base.getindex(layout::RegularArray, i::Int)

Base.getindex — Method

Base.getindex(layout::RegularArray, f::Symbol)

Base.getindex — Method

Base.getindex(layout::RegularArray, r::UnitRange{Int})

Base.getindex — Method

Base.getindex(layout::UnionArray, i::Int)

Base.getindex — Method

Base.getindex(layout::UnionArray, f::Symbol)

Base.getindex — Method

Base.getindex(layout::UnionArray, r::UnitRange{Int})

Base.getindex — Method

Base.getindex(layout::UnmaskedArray, i::Int)

Base.getindex — Method

Base.getindex(layout::UnmaskedArray, f::Symbol)

Base.getindex — Method

Base.getindex(layout::UnmaskedArray, r::UnitRange{Int})

Base.getindex — Method

Base.getindex(
    layout::TupleArray{CONTENTS,BEHAVIOR},
    i::Int,
) where {CONTENTS<:Base.Tuple{Vararg{Content}},BEHAVIOR}

Base.getindex — Method

Base.getindex(
    layout::RecordArray{FIELDS,CONTENTS,BEHAVIOR},
    i::Int,
) where {FIELDS,CONTENTS<:Base.Tuple{Vararg{Content}},BEHAVIOR}

Base.getindex — Method

Base.getindex(
    layout::RecordArray{FIELDS,CONTENTS,BEHAVIOR},
    f::Symbol,
) where {FIELDS,CONTENTS<:Base.Tuple{Vararg{Content}},BEHAVIOR}

Base.getindex — Method

Base.getindex(
    layout::RecordArray{FIELDS,CONTENTS,BEHAVIOR},
    r::UnitRange{Int},
) where {FIELDS,CONTENTS<:Base.Tuple{Vararg{Content}},BEHAVIOR}

Base.getindex — Method

Base.getindex(
    layout::TupleArray{CONTENTS,BEHAVIOR},
    r::UnitRange{Int},
) where {VALUES<:Content,CONTENTS<:Base.Tuple{VALUES},BEHAVIOR}

Base.getindex — Method

Base.getindex(
    layout::RegularArray{PrimitiveArray{UInt8,BUFFER,:byte},:bytestring},
    i::Int,
) where {BUFFER<:AbstractVector{UInt8}}

Base.getindex — Method

Base.getindex(
    layout::RegularArray{PrimitiveArray{UInt8,BUFFER,:char},:string},
    i::Int,
) where {BUFFER<:AbstractVector{UInt8}}

Base.getindex — Method

Base.getindex(
    layout::ListArray{INDEX,PrimitiveArray{UInt8,BUFFER,:byte},:bytestring},
    i::Int,
) where {INDEX<:IndexBig,BUFFER<:AbstractVector{UInt8}}

Base.getindex — Method

Base.getindex(
    layout::ListArray{INDEX,PrimitiveArray{UInt8,BUFFER,:char},:string},
    i::Int,
) where {INDEX<:IndexBig,BUFFER<:AbstractVector{UInt8}}

Base.getindex — Method

Base.getindex(
    layout::ListOffsetArray{INDEX,PrimitiveArray{UInt8,BUFFER,:byte},:bytestring},
    i::Int,
) where {INDEX<:IndexBig,BUFFER<:AbstractVector{UInt8}}

Base.getindex — Method

Base.getindex(
    layout::ListOffsetArray{INDEX,PrimitiveArray{UInt8,BUFFER,:char},:string},
    i::Int,
) where {INDEX<:IndexBig,BUFFER<:AbstractVector{UInt8}}

Base.getindex — Method

Base.getindex(
    layout::SlotRecord{CONTENTS},
    f::Int64,
) where {CONTENTS<:Base.Tuple{Vararg{Content}}}

Base.getindex — Method

Base.getindex(
    layout::Record{FIELDS,CONTENTS},
    f::Symbol,
) where {FIELDS,CONTENTS<:Base.Tuple{Vararg{Content}}}

Base.iterate — Method

Base.iterate(layout::Content, state)

Iteration: The iteration continues by repeatedly calling Base.iterate(layout, state) with the collection and the current state. This returns the next element and the next state until it returns nothing, indicating the end of the iteration.

Parameters

layout::Content: This specifies that the function operates on an instance of the type Content.

state: This represents the current state of the iteration, typically an index or position in the collection.

Base.iterate — Method

Base.iterate(layout::Content)

Enable the use of Julia's iteration protocol on instances of Content type.

Examples

# Assuming Content is defined and an instance is created
for element in layout_instance
    println(element)
end

Initialization: The iteration process starts by calling Base.iterate(layout) with the collection layout as the only argument. This should return the first element and the initial state.

Base.keys — Method

Base.keys(parameters::Parameters)

Base.lastindex — Method

Base.lastindex(layout::BitMaskedArray)

Base.lastindex — Method

Base.lastindex(layout::ByteMaskedArray)

Base.lastindex — Method

Base.lastindex(layout::EmptyArray)

Base.lastindex — Method

Base.lastindex(layout::IndexedArray)

Base.lastindex — Method

Base.lastindex(layout::IndexedOptionArray)

Base.lastindex — Method

Base.lastindex(layout::ListArray)

Last index of a ListArray starts.

Base.lastindex — Method

Base.lastindex(layout::ListOffsetArray)

Last index of a ListOffsetArray offsets.

Base.lastindex — Method

Base.lastindex(layout::PrimitiveArray)

Base.lastindex — Method

Base.lastindex(layout::RecordArray)

Base.lastindex — Method

Base.lastindex(layout::RegularArray)

Last index of a RegularArray.

Base.lastindex — Method

Base.lastindex(layout::TupleArray)

Base.lastindex — Method

Base.lastindex(layout::UnionArray)

Base.lastindex — Method

Base.lastindex(layout::UnmaskedArray)

Base.length — Method

Base.length(layout::BitMaskedArray)

Base.length — Method

Base.length(layout::ByteMaskedArray)

Base.length — Method

Base.length(layout::EmptyArray)

Base.length — Method

Base.length(layout::IndexedArray)

Base.length — Method

Base.length(layout::IndexedOptionArray)

Base.length — Method

Base.length(layout::ListArray)

Length of a ListArray.

Base.length — Method

Base.length(layout::ListOffsetArray)

Length of a ListOffsetArray.

Base.length — Method

Base.length(parameters::Parameters)

Base.length — Method

Base.length(layout::PrimitiveArray)

Base.length — Method

Base.length(layout::RecordArray)

Base.length — Method

Base.length(layout::RegularArray)

Length of a RegularArray.

Base.length — Method

Base.length(layout::TupleArray)

Base.length — Method

Base.length(layout::UnionArray)

Base.length — Method

Base.length(layout::UnmaskedArray)

Base.push! — Method

Base.push!(layout::BitMaskedArray, input)

Base.push! — Method

Base.push!(layout::ByteMaskedArray, input)

Base.push! — Method

Base.push!(layout::EmptyArray, input)

Base.push! — Method

Base.push!(layout::IndexedArray, input)

Base.push! — Method

Base.push!(layout::IndexedOptionArray, input)

Base.push! — Method

Base.push!(layout::ListType, input::AbstractVector)

Base.push! — Method

Base.push!(special::Specialization, input)

Base.push! — Method

Base.push!(layout::UnionArray, input)

Base.push! — Method

Base.push!(layout::UnmaskedArray, input)

Base.push! — Method

Base.push!(layout::ListType{BEHAVIOR}, input::String) where {BEHAVIOR}

Base.push! — Method

Base.push!(
    layout::RecordArray{FIELDS,CONTENTS},
    input::NamedTuple{FIELDS},
) where {FIELDS,CONTENTS<:Base.Tuple{Vararg{Content}}}

Base.push! — Method

Base.push!(
    layout::TupleArray{CONTENTS},
    input::INPUT,
) where {N,CONTENTS<:Base.Tuple{Vararg{Content,N}},INPUT<:Base.Tuple{Vararg{Any,N}}}

Base.push! — Method

Base.push!(layout::PrimitiveArray{ITEM}, input::ITEM) where {ITEM}

Push multiple ITEM elements (from another collection) to an instance of PrimitiveArray.

Base.push! — Method

Base.push!(layout::PrimitiveArray{ITEM}, input::Number) where {ITEM}

Base.show — Method

Base.show(io::IO, parameters::Parameters)

Base.show — Method

Base.show(
    io::IO,
    data::Union{Content,Record,Tuple};
    limit_rows::Int = 1,
    limit_cols::Int = 80,
)

Base.show — Method

Base.show(data::Union{Content,Record,Tuple}; limit_rows::Int = 1, limit_cols::Int = 80)

Base.size — Method

Treat instances of Content as if they are one-dimensional arrays.

Index

AwkwardArray.ByteStringArray

AwkwardArray.ByteStringArray

AwkwardArray.ByteStringOffsetArray

AwkwardArray.ByteStringOffsetArray

AwkwardArray.ByteStringRegularArray

AwkwardArray.ByteStringRegularArray

AwkwardArray.ByteStringRegularArray

AwkwardArray.StringArray

AwkwardArray.StringArray

AwkwardArray.StringArray

AwkwardArray.StringOffsetArray

AwkwardArray.StringOffsetArray

AwkwardArray.StringOffsetArray

AwkwardArray.StringRegularArray

AwkwardArray.StringRegularArray

AwkwardArray.StringRegularArray

AwkwardArray.StringRegularArray

AwkwardArray._alternate

AwkwardArray._get_buffer

AwkwardArray._get_index

AwkwardArray._horizontal

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers!

AwkwardArray._to_buffers_index

AwkwardArray._to_buffers_parameters

AwkwardArray._vertical

AwkwardArray.check_primitive_type

AwkwardArray.compatible

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.copy

AwkwardArray.default_buffer_key

AwkwardArray.end_list!

AwkwardArray.end_list!

AwkwardArray.end_list!

AwkwardArray.end_list!

AwkwardArray.end_list!

AwkwardArray.end_list!

AwkwardArray.end_list!

AwkwardArray.end_list!

AwkwardArray.end_list!

AwkwardArray.end_record!

AwkwardArray.end_record!

AwkwardArray.end_record!

AwkwardArray.end_record!

AwkwardArray.end_record!

AwkwardArray.end_record!

AwkwardArray.end_record!

AwkwardArray.end_tuple!

AwkwardArray.end_tuple!

AwkwardArray.end_tuple!

AwkwardArray.end_tuple!

AwkwardArray.end_tuple!

AwkwardArray.end_tuple!

AwkwardArray.end_tuple!

AwkwardArray.from_buffers

AwkwardArray.from_buffers

AwkwardArray.from_iter

AwkwardArray.get_parameter

AwkwardArray.get_parameter

AwkwardArray.get_parameter

AwkwardArray.has_parameter

AwkwardArray.has_parameter

AwkwardArray.has_parameter

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.is_valid

AwkwardArray.isprimitive

AwkwardArray.layout_for

AwkwardArray.parameters_of

AwkwardArray.parameters_of

AwkwardArray.push_dummy!

AwkwardArray.push_dummy!

AwkwardArray.push_dummy!

AwkwardArray.push_dummy!

AwkwardArray.push_dummy!

AwkwardArray.push_dummy!

AwkwardArray.push_dummy!

AwkwardArray.push_dummy!

AwkwardArray.push_dummy!

AwkwardArray.push_dummy!

AwkwardArray.push_dummy!

AwkwardArray.push_dummy!

AwkwardArray.push_null!

AwkwardArray.push_null!

AwkwardArray.push_null!

AwkwardArray.push_null!

AwkwardArray.slot

AwkwardArray.slot

AwkwardArray.to_buffers

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector

AwkwardArray.to_vector_or_scalar

AwkwardArray.to_vector_or_scalar

AwkwardArray.with_parameter

AwkwardArray.with_parameter

Base.:==

Base.:==

Base.:==

Base.:==

Base.:==

Base.append!

Base.append!

Base.append!

Base.eltype

Base.eltype

Base.eltype

Base.eltype

Base.eltype

Base.eltype

Base.eltype

Base.eltype

Base.eltype

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.firstindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.getindex

Base.iterate

Base.iterate

Base.keys

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.lastindex

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.length

Base.push!

Base.push!

Base.push!

Base.push!

Base.push!

Base.push!

Base.push!

Base.push!

Base.push!

Base.push!

Base.push!

Base.push!

Base.push!

Base.push!

Base.show

Base.show

Base.show

Base.size